refactor of task3.2 - get square of Shape
