+++ src/main/java/geometry/Rectangle.java
@@ -3,6 +3,7 @@ package geometry;
 public class Rectangle extends Rectangular {
 
     public Rectangle(double sideA, double sideB) {
+
         super(sideA, sideB);
     }
