+++ src/main/java/geometry/Round.java
@@ -5,13 +5,8 @@ public abstract class Round implements Shape {
     // special manipulation to have correct PI 3.14159
     public static final double PI = Math.round(Math.PI * Math.pow(10, 5)) / Math.pow(10, 5);
 
-    private double radius;
     private double axisA;
     private double axisB;
-
-    public Round(double radius) {
-        this.radius = radius;
-    }
 
     public Round(double axisA, double axisB) {
         this.axisA = axisA;
@@ -27,7 +22,7 @@ public abstract class Round implements Shape {
     }
 
     public double getSquare() {
-        return PI * Math.pow(this.radius, 2);
+        return PI * axisA * axisB;
     }
 
 }
